requirements.txt
python-telegram-bot
requests
ffmpeg-python
